mcp-servers/truenas-mcp-server/requirements.txt
mcp>=1.1.0
httpx>=0.27.0
python-dotenv>=1.0.0